+++ bpdm-common-test/src/main/kotlin/org/eclipse/tractusx/bpdm/test/testdata/orchestrator/OrchestratorRequestFactoryCommon.kt
@@ -224,35 +224,26 @@ class OrchestratorRequestFactoryCommon(
 
     fun buildPhysicalAddressScriptVariant(scriptCode: String, seed: String): PhysicalAddressScriptVariant{
         return PhysicalAddressScriptVariant(
-            postalCode = buildScriptVariantStringValue("Postal Code", seed, scriptCode),
             city = buildScriptVariantStringValue("City", seed, scriptCode),
             district = buildScriptVariantStringValue("District", seed, scriptCode),
-            street = Street(
+            street = StreetScriptVariant(
                 name = buildScriptVariantStringValue("Street Name", seed, scriptCode),
-                houseNumber = buildScriptVariantStringValue("HouseNumber", seed, scriptCode),
-                houseNumberSupplement =  buildScriptVariantStringValue("House Number Supplement", seed, scriptCode),
-                milestone = buildScriptVariantStringValue("Milestone", seed, scriptCode),
                 direction = buildScriptVariantStringValue("Direction", seed, scriptCode),
                 namePrefix = buildScriptVariantStringValue("NamePrefix", seed, scriptCode),
                 additionalNamePrefix = buildScriptVariantStringValue("AdditionalNamePrefix", seed, scriptCode),
                 nameSuffix = buildScriptVariantStringValue("NameSuffix", seed, scriptCode),
                 additionalNameSuffix = buildScriptVariantStringValue("AdditionalNameSuffix", seed, scriptCode),
             ),
-            companyPostalCode = buildScriptVariantStringValue("CompanyPostal Code", seed, scriptCode),
             industrialZone = buildScriptVariantStringValue("IndustrialZone", seed, scriptCode),
             building = buildScriptVariantStringValue("Building", seed, scriptCode),
             floor = buildScriptVariantStringValue("Floor", seed, scriptCode),
             door = buildScriptVariantStringValue("Door", seed, scriptCode),
-            taxJurisdictionCode = buildScriptVariantStringValue("Tax Jurisdiction", seed, scriptCode),
         )
     }
 
     fun buildAlternativeAddressScriptVariant(scriptCode: String, seed: String): AlternativeAddressScriptVariant{
         return AlternativeAddressScriptVariant(
-            postalCode = buildScriptVariantStringValue("Postal Code $seed", seed, scriptCode),
-            city = buildScriptVariantStringValue("City $seed", seed, scriptCode),
-            deliveryServiceQualifier = buildScriptVariantStringValue("Delivery Service Qualifier $seed", seed, scriptCode),
-            deliveryServiceNumber = buildScriptVariantStringValue("Delivery Service Number $seed", seed, scriptCode)
+            city = buildScriptVariantStringValue("City $seed", seed, scriptCode)
         )
     }