+++ bpdm-gate/src/main/kotlin/org/eclipse/tractusx/bpdm/gate/service/BusinessPartnerMappings.kt
@@ -266,16 +266,13 @@ class BusinessPartnerMappings {
     private fun toPhysicalAddressScriptVariantDto(entity: PhysicalPostalAddressScriptVariantDb): PhysicalAddressScriptVariantDto{
         return with(entity) {
             PhysicalAddressScriptVariantDto(
-                postalCode = postalCode,
                 city = city,
                 district = district,
-                street = street?.toStreetDto() ?: StreetDto(),
-                companyPostalCode = companyPostalCode,
+                street = street?.toStreetScriptVariantDto() ?: StreetScriptVariantDto(),
                 industrialZone = industrialZone,
                 building = building,
                 floor = floor,
-                door = door,
-                taxJurisdictionCode = taxJurisdictionCode
+                door = door
             )
         }
     }
@@ -283,10 +280,7 @@ class BusinessPartnerMappings {
     private fun toAlternativeAddressScriptVariantDto(entity: AlternativePostalAddressScriptVariantDb): AlternativeAddressScriptVariantDto{
         return with(entity){
             AlternativeAddressScriptVariantDto(
-                postalCode = postalCode,
-                city = city,
-                deliveryServiceQualifier = deliveryServiceQualifier,
-                deliveryServiceNumber = deliveryServiceNumber
+                city = city
             )
         }
     }
@@ -309,16 +303,13 @@ class BusinessPartnerMappings {
     private fun toPhysicalAddressScriptVariantDb(dto: PhysicalAddressScriptVariantDto): PhysicalPostalAddressScriptVariantDb{
         return with(dto){
             PhysicalPostalAddressScriptVariantDb(
-                postalCode = postalCode,
                 city = city,
                 district = district,
-                street = toStreet(street),
-                companyPostalCode = companyPostalCode,
+                street = toStreetScriptVariant(street),
                 industrialZone = industrialZone,
                 building = building,
                 floor = floor,
-                door = door,
-                taxJurisdictionCode = taxJurisdictionCode
+                door = door
             )
         }
     }
@@ -326,12 +317,33 @@ class BusinessPartnerMappings {
     private fun toAlternativeAddressScriptVariantDb(dto: AlternativeAddressScriptVariantDto): AlternativePostalAddressScriptVariantDb{
         return with(dto){
             AlternativePostalAddressScriptVariantDb(
-                postalCode = postalCode,
-                city = city,
-                deliveryServiceQualifier = deliveryServiceQualifier,
-                deliveryServiceNumber = deliveryServiceNumber
+                city = city
             )
         }
+    }
+
+    private fun toStreetScriptVariant(dto: StreetScriptVariantDto): StreetScriptVariantDb{
+        return with(dto){
+            StreetScriptVariantDb(
+                name = name,
+                direction = direction,
+                namePrefix = namePrefix,
+                additionalNamePrefix = additionalNamePrefix,
+                nameSuffix = nameSuffix,
+                additionalNameSuffix = additionalNameSuffix
+            )
+        }
+    }
+
+    private fun StreetScriptVariantDb.toStreetScriptVariantDto(): StreetScriptVariantDto{
+        return StreetScriptVariantDto(
+            name = name,
+            direction = direction,
+            namePrefix = namePrefix,
+            additionalNamePrefix = additionalNamePrefix,
+            nameSuffix = nameSuffix,
+            additionalNameSuffix = additionalNameSuffix
+        )
     }
 
     // convert empty DTO to null