+++ bpdm-gate/src/main/kotlin/org/eclipse/tractusx/bpdm/gate/service/GoldenRecordUpdateServices.kt
@@ -553,24 +553,28 @@ class GoldenRecordUpdateChunkService(
 
     private fun PhysicalAddressScriptVariantDto.toEntity() =
         PhysicalPostalAddressScriptVariantDb(
-            postalCode = postalCode,
             city = city,
             district = district,
             street = street?.toEntity(),
-            companyPostalCode = companyPostalCode,
             industrialZone = industrialZone,
             building = building,
             floor = floor,
-            door = door,
-            taxJurisdictionCode = taxJurisdictionCode
+            door = door
         )
 
     private fun AlternativeAddressScriptVariantDto.toEntity() =
         AlternativePostalAddressScriptVariantDb(
-            postalCode = postalCode,
-            city = city,
-            deliveryServiceQualifier = deliveryServiceQualifier,
-            deliveryServiceNumber = deliveryServiceNumber
+            city = city
+        )
+
+    private fun StreetScriptVariantDto.toEntity() =
+        StreetScriptVariantDb(
+            name,
+            direction,
+            namePrefix,
+            additionalNamePrefix,
+            nameSuffix,
+            additionalNameSuffix
         )